consolidate services into taskService and listService
