--- src/services/getListService.js
@@ -1,11 +1,0 @@
-const { stringify } = require('flatted');
-const todoListRepository = require('../entities/todoListRepository');
-
-const getListService = (todoListId) => {
-  const todoLists = todoListRepository.getTodoListById(todoListId);
-  return stringify(todoLists);
-};
-
-module.exports = {
-  getListService,
-};
