--- src/services/getListsService.js
@@ -1,11 +1,0 @@
-const { stringify } = require('flatted');
-const todoListRepository = require('../entities/todoListRepository');
-
-const getListsService = () => {
-  const { todoLists } = todoListRepository;
-  return stringify(todoLists);
-};
-
-module.exports = {
-  getListsService,
-};
